backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
sqlalchemy
pymysql
aiomysql
pytest
pytest-asyncio
httpx

openai